+++ src/main/java/com/example/api/possbileboard/dto/PossibleDetails.java
@@ -11,6 +11,6 @@ public record PossibleDetails(
         LocalDateTime possibleStartTime,
         LocalDateTime possibleEndTime,
         Long contractCount,
-        Integer starPoint
+        Double starPoint
 ) {
 }
